% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhds-package.R
\docType{package}
\name{nhds-package}
\alias{nhds}
\alias{nhds-package}
\title{nhds: National Hospital Discharge Survey 2010 Data}
\description{
The National Hospital Discharge Survey (2010)
    summarizes the state of patient at the end of hospital admissions for
    that year. The data is from a subset of US hospitals and is
    deidentified.
}
\author{
\strong{Maintainer}: Jack Wasey \email{first.last@example.com}

}
\keyword{internal}
