% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhds-package.R
\docType{data}
\name{nhds2010}
\alias{nhds2010}
\title{National Hospital Discharge Survey 2010 data}
\source{
\url{https://www.cdc.gov/nchs/nhds/index.htm}
}
\description{
This is public domain data from the US \href{https://www.cdc.gov}{Center for Disease Control}. This is deidentified patient data describing
the information at the end of a hospital admission, including demographic
information, admission diagnoses, comorbidities and procedure codes, death or
disposition. There are no identifiers in the data, so an integer count was
included in this data for this purpose.
}
\keyword{datasets}
