% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhds.R
\name{parse_nhds2010}
\alias{parse_nhds2010}
\title{Parse the National Hospital Discharge Survey data}
\usage{
parse_nhds2010(save = TRUE)
}
\arguments{
\item{save}{logical, if `TRUE` saves data in the `data/` directory in the
package source tree}
}
\description{
The raw data files are large, so are not included in the git repository or
package. The raw data is 22M, and the R data is 3.3M when compressed. The
data can be downloaded from the referenced URL.
}
\examples{
library(icd)
colSums(comorbid_charlson(nhds2010))
}
\references{
\url{https://www.cdc.gov/nchs/nhds/index.htm}
}
\keyword{internal}
