% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhds-package.R
\docType{package}
\name{nhds-package}
\alias{nhds}
\alias{nhds-package}
\title{nhds: National Hospital Discharge Survey 2010 Data}
\description{
The National Hospital Discharge Survey (2010)
    summarizes the state of patient at the end of hospital admissions for
    that year. The data is from a subset of US hospitals and is
    deidentified.
}
\examples{
nhds2010[1:5, ]
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jackwasey/nhds}
  \item Report bugs at \url{https://github.com/jackwasey/nhds/issues}
}

}
\author{
\strong{Maintainer}: Jack Wasey \email{jack@jackwasey.com}

}
\keyword{internal}
